Refactor dashboard route and add image type declarations
Moved ModuleCreatorDashboard to a new directory and updated its route to '/modulecreator-dashboard' in App and sidebar components. Added a declarations.d.ts file for image module types to improve TypeScript support. Updated @testing-library/jest-dom to version 6.7.0 in package files.
