+++ web-client/src/declarations.d.ts
@@ -1,0 +1,5 @@
+declare module "*.gif";
+declare module "*.png";
+declare module "*.jpg";
+declare module "*.jpeg";
+declare module "*.svg";
